move files around and change paths
