+++ aggregator/src/aggregation.rs
@@ -1,9 +1,5 @@
-/// Config to evaluate blob polynomial at a random challenge.
-mod barycentric;
 /// Config to constrain batch data (decoded blob data)
 mod batch_data;
-/// Config to constrain blob data (encoded batch data)
-mod blob_data;
 /// Circuit implementation of aggregation circuit.
 mod circuit;
 /// Config for aggregation circuit
@@ -15,13 +11,9 @@ mod rlc;
 /// Utility module
 mod util;
 
-pub(crate) use barycentric::{
-    interpolate, AssignedBarycentricEvaluationConfig, BarycentricEvaluationConfig, BLS_MODULUS,
-};
 pub(crate) use batch_data::BatchDataConfig;
-pub(crate) use blob_data::BlobDataConfig;
 pub(crate) use decoder::{witgen, DecoderConfig, DecoderConfigArgs};
-pub(crate) use rlc::RlcConfig;
+pub(crate) use rlc::{RlcConfig, POWS_OF_256};
 
 pub use circuit::BatchCircuit;
 pub use config::BatchCircuitConfig;
