+++ aggregator/src/data_availability.rs
@@ -6,16 +6,16 @@ enum DataAvailibility {
 }
 
 cfg_if! {
-    if #[cfg(feature = "da-4844")] {
+    if #[cfg(feature = "da-eip4844")] {
         const DATA_AVAILABILITY: DataAvailibility = DataAvailibility::Eip4844;
     } else if #[cfg(feature = "da-avail")] {
         const DATA_AVAILABILITY: DataAvailibility = DataAvailibility::Avail;
     } else {
-        unimplemented!("no da feature flag set");
+        compile_error!("no da feature flag set");
     }
 }
 
-#[cfg(feature = "da-4844")]
-mod eip4844;
 #[cfg(feature = "da-avail")]
 mod avail;
+#[cfg(feature = "da-eip4844")]
+pub mod eip4844;
