+++ aggregator/src/lib.rs
@@ -4,8 +4,6 @@ mod aggregation;
 /// This module implements `Batch` related data types.
 /// A batch is a list of chunk.
 mod batch;
-/// blob struct and constants
-mod blob;
 /// Config to recursive aggregate multiple aggregations
 mod recursion;
 // This module implements `Chunk` related data types.
@@ -17,14 +15,12 @@ mod compression;
 mod constants;
 /// Core module for circuit assignment
 mod core;
-/// EIP-4844 related utils.
-pub mod eip4844;
+/// Data availibility checks
+pub mod data_availability;
 /// Parameters for compression circuit
 mod param;
 /// utilities
 mod util;
-/// Data availibility checks
-mod data_availability;
 
 #[cfg(test)]
 mod tests;
@@ -32,10 +28,10 @@ mod tests;
 pub use self::core::extract_proof_and_instances_with_pairing_check;
 pub use aggregation::*;
 pub use batch::{BatchHash, BatchHeader};
-pub use blob::BatchData;
 pub use chunk::ChunkInfo;
 pub use compression::*;
 pub use constants::MAX_AGG_SNARKS;
 pub(crate) use constants::*;
+pub use data_availability::eip4844::blob::BatchData;
 pub use param::*;
 pub use recursion::*;
