+++ prover/src/aggregator.rs
@@ -2,5 +2,5 @@ mod prover;
 mod verifier;
 
 pub use self::prover::{check_chunk_hashes, Prover};
-pub use aggregator::{eip4844, BatchData, BatchHash, BatchHeader, MAX_AGG_SNARKS};
+pub use aggregator::{BatchData, BatchHash, BatchHeader, MAX_AGG_SNARKS};
 pub use verifier::Verifier;
